\name{gQTLstats-package}
\alias{gQTLstats-package}
\alias{gQTLstats}
\alias{prep.cisAssocNB} 
\alias{storeToMaxAssocBySNP} 
\alias{table_sensobj_thresh}

\docType{package}
\title{
\packageTitle{gQTLstats}
}
\description{
\packageDescription{gQTLstats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gQTLstats}
\packageIndices{gQTLstats}

This package addresses the management of map-reduce like computations
for cis-association tests between DNA variants and genomic features
like gene expression measurements.  It makes essential use of
data structures defined in package gQTLBase.

A number of experimental functions are present in the 
current version of the package:
prep.cisAssocNB (assembles information to assess negative
binomial regression in cis association testing),
storeToMaxAssocBySNP (progress towards SNP-specific FDR),
table_sensobj_thresh (reporting on sensitivity analysis).
}
\author{
\packageAuthor{gQTLstats}

Maintainer: \packageMaintainer{gQTLstats}
}
\keyword{ package }
%\examples{
%~~ simple examples of the most important functions ~~
%}
